correction des erreurs

+++ src/Arc.java
@@ -1,8 +1,15 @@
+import java.util.List;
+import java.util.ArrayList;
+
+
+
+
+
 public class Arc {
     private double poids;
     private String cible;
 
-    public Arc (int p, String s){
+    public Arc (double p, String s){
         this.poids = p;
         this.cible = s;
     }
